API: Move TyphonTool to top level directory rename TyphonBase
