+++ typhon/tools/__init__.py
@@ -1,6 +1,5 @@
 """Module for all insertable Typhon tools"""
-__all__ = ['TyphonLogDisplay', 'TyphonTimePlot', 'TyphonTool']
+__all__ = ['TyphonLogDisplay', 'TyphonTimePlot']
 
-from .core import TyphonTool
 from .plot import TyphonTimePlot
 from .log import TyphonLogDisplay
